requirements.txt
Django~=3.0.7
python-dotenv==0.13.0
gunicorn==20.0.4
whitenoise==5.1.0
pillow==7.1.2
django-taggit==1.3.0
